AMP-61335 code gen clean up
